email-agent/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.3.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
sse-starlette>=2.0.0

# Optional: Agent Skeleton production backends
# redis>=5.0.0                         # RedisStateBackend
# opentelemetry-api>=1.20.0            # TracedAgentRunner tracing
# opentelemetry-sdk>=1.20.0            # TracedAgentRunner tracing
